+++ packages/frontend/src/pages/users/[userId]/apitokens.vue
@@ -29,10 +29,22 @@
 
 <script setup lang="ts">
 const route = useRoute();
-const apiTokens = await useFetchE<ApiToken[]>(`/api/v1/pentestusers/${route.params.userId}/apitokens/`, { method: 'GET' });
+const auth = useAuth();
+const apiTokens = await useAsyncDataE<ApiToken[]>(async () => {
+  try {
+    return await $fetch<ApiToken[]>(`/api/v1/pentestusers/${route.params.userId}/apitokens/`, { method: 'GET' });
+  } catch (error: any) {
+    if (error?.data?.code === 'reauth-required') {
+      auth.redirectToReAuth({ replace: true });
+      return [];
+    } else {
+      throw error;
+    }
+  }
+});
 
 async function deleteApiToken(apiToken: ApiToken) {
-  await $fetch(`/api/v1/pentestusers/self/apitokens/${apiToken.id}/`, { method: 'GET' });
+  await $fetch(`/api/v1/pentestusers/${route.params.userId}/apitokens/${apiToken.id}/`, { method: 'DELETE' });
   apiTokens.value = apiTokens.value.filter(t => t.id !== apiToken.id);
 }
 </script>